[REFACTOR] [#1] Survey 관련 클래스 패키지 구조 변경
값 타입 클래스 value 패키지로 위치 변경
